journey_service/requirements.txt
amqp==5.3.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.10.0
async-timeout==5.0.1
Authlib==1.6.5
billiard==4.2.2
cachetools==6.2.1
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
cryptography==46.0.3
decorator==5.2.1
Django==5.2.8
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
future==1.0.0
geocoder==1.38.1
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
gunicorn==23.0.0
httplib2==0.31.0
idna==3.11
jwt==1.4.0
kafka-python==2.2.15
kombu==5.5.4
oauthlib==3.3.1
packaging==25.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyparsing==3.2.5
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
ratelim==0.1.6
redis==7.0.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
